+++ 1-grid-world/2-value_iteration.py
@@ -1,4 +1,4 @@
-from gridworld import GraphicDisplay, PolicyEnv as Env
+from env import GraphicDisplay, PolicyEnv as Env
 
 
 class ValueIteration:
