+++ 1-grid-world/4-q_learning.py
@@ -2,7 +2,7 @@ import numpy as np
 import random
 from collections import defaultdict
 
-from gridworld import Env
+from env import Env
 
 class QLearningAgent:
     def __init__(self, actions):
