+++ 1-grid-world/5-deep_sarsa.py
@@ -20,7 +20,7 @@ import torch
 import torch.nn as nn
 import torch.optim as optim
 
-from gridworld import DynamicEnv
+from env import DynamicEnv
 
 EPISODES = 1000
